--- apps/blog/src/app/posts/[id]/_components/delete-post-modal.tsx
@@ -1,65 +1,0 @@
-'use client';
-
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from '@nettee-sample/ui/components/alert-dialog';
-import { Button } from '@nettee-sample/ui/components/button';
-import { useRouter } from 'next/navigation';
-import { useState } from 'react';
-
-import { postAPI } from '@/lib/api';
-import { ROUTES } from '@/lib/constants/routes';
-
-interface Props {
-  postId: string;
-}
-
-export default function DeletePostModal({ postId }: Props) {
-  const router = useRouter();
-  const [isDeleting, setIsDeleting] = useState(false);
-
-  const handleDelete = async () => {
-    if (isDeleting) return;
-
-    setIsDeleting(true);
-    try {
-      await postAPI.delete(postId);
-      router.push(ROUTES.home());
-    } catch (error) {
-      console.error('Failed to delete post:', error);
-    } finally {
-      setIsDeleting(false);
-    }
-  };
-
-  return (
-    <AlertDialog>
-      <AlertDialogTrigger asChild>
-        <Button variant="destructive">삭제</Button>
-      </AlertDialogTrigger>
-      <AlertDialogContent>
-        <AlertDialogHeader>
-          <AlertDialogTitle>게시글 삭제</AlertDialogTitle>
-          <AlertDialogDescription>이 게시글을 삭제하시겠습니까? 이 작업은 되돌릴 수 없습니다.</AlertDialogDescription>
-        </AlertDialogHeader>
-        <AlertDialogFooter>
-          <AlertDialogCancel>취소</AlertDialogCancel>
-          <AlertDialogAction
-            className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-            disabled={isDeleting}
-            onClick={handleDelete}>
-            {isDeleting ? '삭제 중...' : '삭제'}
-          </AlertDialogAction>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
-  );
-}
