+++ apps/blog/src/app/posts/[id]/edit/page.tsx
@@ -1,8 +1,7 @@
 import { notFound } from 'next/navigation';
 
-import { postAPI } from '@/lib/api';
-
-import EditForm from './_components/post-edit-form';
+import EditPostForm from '@/features/post/edit-post/edit-post.ui';
+import { postAPI } from '@/shared/api';
 
 interface Props {
   params: Promise<{
@@ -17,7 +16,7 @@ export default async function EditPage({ params }: Props) {
     return (
       <div>
         <h1 className="mb-6 text-3xl font-bold">새 게시글 작성</h1>
-        <EditForm post={post} />
+        <EditPostForm post={post} />
       </div>
     );
   } catch (error) {
