--- apps/blog/src/components/providers.tsx
@@ -1,19 +1,0 @@
-'use client';
-
-import { ThemeProvider as NextThemesProvider } from 'next-themes';
-
-import WriteForm from '@/app/posts/write/_components/post-form';
-
-export function Providers({ children }: { children: React.ReactNode }) {
-  return (
-    <NextThemesProvider
-      attribute="class"
-      defaultTheme="system"
-      disableTransitionOnChange
-      enableColorScheme
-      enableSystem>
-      <WriteForm />
-      {children}
-    </NextThemesProvider>
-  );
-}
